Share common code among apis (including backend) via kpm-api-common
