+++ apis/my-canvas-rooms-api/package.json
@@ -24,6 +24,7 @@
     "@kth/canvas-api": "^4.2.2",
     "dotenv": "^16.0.2",
     "express": "^4.18.1",
+    "kpm-api-common": "^0.1.0",
     "ts-node": "^10.9.1"
   }
 }
