+++ kpm/kpm-backend/package.json
@@ -21,6 +21,7 @@
     "dotenv": "^16.0.2",
     "express": "^4.18.1",
     "got": "^11.8.5",
+    "kpm-api-common": "^0.1.0",
     "ts-node": "^10.9.1"
   }
 }
